chore(cleanup): rename pages/categorie to pages/category
