+++ src/gui/pages/category/categoriepage.py
